+++ nodes/ui/mantine/src/molecules/tableSelectionScope/tableSelectionScopeNode.ts
@@ -16,7 +16,24 @@ export default reactNode('TableSelectionScope', {
         ],
         outputs: [
             getPort({ plug: 'output', name: 'selectionScope', displayName: 'selectionScope', group: 'Data', type: 'object' }),
-            getPort({ plug: 'output', name: 'selectionByTableId', displayName: 'selectionByTableId', group: 'Data', type: 'object' }),
+            getPort({ plug: 'output', name: 'selectionByBDClass', displayName: 'selectionByBDClass', group: 'Data', type: 'object' }),
+            getPort({ plug: 'output', name: 'changed', displayName: 'changed', group: 'Signals', type: 'signal' }),
+        ]
+    },
+    'v1.0.1': {
+        module: {
+            dynamic: lazy(() => import(
+                /* webpackPrefetch: true */
+                /* webpackPreload: true */
+                '@packages/table-selection-scope-v1.0.1'))
+        },
+        inputs: [
+            ...getPorts('input', ['reset']),
+            getPort({ plug: 'input', name: 'newSelectionScope', displayName: 'selectionScope', group: 'Data', type: 'object' }),
+        ],
+        outputs: [
+            getPort({ plug: 'output', name: 'selectionScope', displayName: 'selectionScope', group: 'Data', type: 'object' }),
+            getPort({ plug: 'output', name: 'selectionByBDClass', displayName: 'selectionByBDClass', group: 'Data', type: 'object' }),
             getPort({ plug: 'output', name: 'changed', displayName: 'changed', group: 'Signals', type: 'signal' }),
         ]
     }